Editor, buckets and file navigator (#38)
* Hooks must be used within their providers

* Contexts should get locale from the req props

* Move FileNavigatorItem to a type file

* Create bucket transform utilities

* Bucket delete endpoints should also return response

* transformItem bucket can be null

* Remove stored bucket from the bucket service

* Create BucketContext.tsx

* Update bucket types according to the API

* Update FileNavigatorItem to work with the updated Bucket logic

* Update bucket utils to match type

* Bucket service should use FileNavigatorItem type with originals and current versions

* Update bucket context functions

* Rename: Action editor -> Editor

* Add type to FileNavigatorItem (folder or file)

* Update BucketService.ts

* Refactor file navigator

* Move the fileNavigator to editor

* Create Workspace layout which uses BucketProvider

* Remove inSidebar property from tables

* Create editor layout and wrapper

* Update naming and links for lists

* Remove /viewer paths

* Move MDX Editor to /editor

* Content -> Contents

* content -> contents

* Move widgets and creation form to /dashboards component dir

* Bucket files and folders will not have a parent

* Delete actionResultsAndTabs.tsx

* Fix typo in folder name

* Create editorResults and datasetViewer components

* Clean up

* Remove files import from editor page

* Update BucketContext.tsx

* Manage list of open editor files in the bucket context

* Confirm deletions

* updateStateWithBucket util

* Editor wip

* Fix lint

* Fix table placeholder data

* UI fixes for the editor

* Move activetab to the context

* Add the formatted table name to the editor when selected

* Stylisting changes to the editor

* Make the top bar smaller

* Move logo to the top bar, when

* Make auth screen inputs solid

* Update workspace switcher page

* Move workspace management to it's own component

* Use @todo instead of TODO in tsdoc

* Remove theme from docs

* Fix docs link

* Fix Bucket transformer

* Add missing translations

* Fail safes for bucket fetching

* Small UI changes

* Create and implement addNew modal

* Update PopupContext.tsx

* Implement name formatting when creating

* Add creation validation

* Change the way newItemData is created

* Improve type checks

* Add button for context menu

* Add move option to context menu

* Move updateFieldValues to bucketUtils

* UI to handle rename or move item

* Fix context menu UI

* When folder is changed, all children are affected

* Close context menu on action

* Set the local state on update

* Update the whole state after updateFieldValues

* Destructure props for updateFieldValues

* Update translations

* Refactor confirm popup to stop triggering success function instantly

* Add docs to popups

* Implement update item functions in RenameOrMoveItemModal

* Fix docs

* Move confirmations from context to component, handle delete

* Fix type confusion

* Close tabs on edit in file nav

* Reworking the editor flow

* Create SaveEditorAsFileModalContent

* Implement modals on editor

* Fix typo

* Update editorResults.tsx

* Fix website nav

* Remove useless logs

* Update examples

* Update websiteNavigationContent.tsx

* Update websiteNavigationContent.tsx

* Fix Set the content for the open tabs

* Fix editor result max height

* Update BucketContext.tsx

* Remove eslint-plugin-tsdoc

* Add sharp for Image Optimization

https://nextjs.org/docs/messages/sharp-missing-in-production

* Update portalNavProfile.tsx

* Rework portal sidebar useful links

* Add more useful links

* Bottom links are now useful links

* Move buttons to match the white bar

* Update portalNavigation.tsx

* Determine the width of the content based on the sidebar

* Debounce the resize observer

* Transitions and debounce

* 81:45  Warning: The ref value 'sidebarRef.current' will likely have changed by the time this effect cleanup function runs. If this ref points to a node rendered by React, copy 'sidebarRef.current' to a variable inside the effect, and use that variable in the cleanup function.  react-hooks/exhaustive-deps

* Update README.md

* Update editorNew.tsx

* Remove logs

* Update RenameOrMoveItemModalContent.tsx

* Translate error modals

* Translate alert titles

* Maybe path is null

* Update README.md

+++ irmin-console/src/components/workspaceUsersAndPermissions.tsx
@@ -168,28 +168,27 @@ const WorkspaceUsersAndPermissions: React.FC = () => {
     irminConfirm(
       'info',
       dict.usersPermissions.removeUserConfirmation,
-      async () => {
-        // Removal confirmed
-        try {
-          // Remove user from workspace
-          const res = await deleteUser(id);
-          // Remove user from the list
-          setUsers(users.filter((user) => user.id !== id));
-          irminAlert(
-            'success',
-            res.metadata?.message ??
-              'User removed successfully from the workspace'
-          );
-        } catch (error) {
-          console.error('Error changing user role:', error);
-          irminAlert(
-            'error',
-            (error as Error)?.message ?? 'Error removing user'
-          );
+      async (confirmed) => {
+        if (confirmed) {
+          // Removal confirmed
+          try {
+            // Remove user from workspace
+            const res = await deleteUser(id);
+            // Remove user from the list
+            setUsers(users.filter((user) => user.id !== id));
+            irminAlert(
+              'success',
+              res.metadata?.message ??
+                'User removed successfully from the workspace'
+            );
+          } catch (error) {
+            console.error('Error changing user role:', error);
+            irminAlert(
+              'error',
+              (error as Error)?.message ?? 'Error removing user'
+            );
+          }
         }
-      },
-      () => {
-        // Removal not confirmed
       }
     );
   };
@@ -197,31 +196,30 @@ const WorkspaceUsersAndPermissions: React.FC = () => {
   const handleTransferOwnership = async (id: number) => {
     // Confirm transfer
     irminConfirm(
-      'info',
+      'warning',
       dict.usersPermissions.transferOwnershipConfirmation,
-      async () => {
-        // Transfer confirmed
-        if (!currentWorkspace) return;
-        try {
-          // Transfer ownership
-          const res = await transferOwnership(id);
-          // Refetch the current workspace
-          switchToWorkspace(currentWorkspace.slug);
-          // Inform that ownership has been transferred
-          irminAlert(
-            'success',
-            res.metadata?.message ?? 'Ownership transfered successfully'
-          );
-        } catch (error) {
-          console.error('Error transferring ownership:', error);
-          irminAlert(
-            'error',
-            (error as Error)?.message ?? 'Error transferring ownership'
-          );
+      async (confirmed) => {
+        if (confirmed) {
+          // Transfer confirmed
+          if (!currentWorkspace) return;
+          try {
+            // Transfer ownership
+            const res = await transferOwnership(id);
+            // Refetch the current workspace
+            switchToWorkspace(currentWorkspace.slug);
+            // Inform that ownership has been transferred
+            irminAlert(
+              'success',
+              res.metadata?.message ?? 'Ownership transfered successfully'
+            );
+          } catch (error) {
+            console.error('Error transferring ownership:', error);
+            irminAlert(
+              'error',
+              (error as Error)?.message ?? 'Error transferring ownership'
+            );
+          }
         }
-      },
-      () => {
-        // Transfer not confirmed
       }
     );
   };
